Create middleware/authentication folder to handle jwt verification
